nfcc_pc/requirements.txt
websockets>=12.0
pystray>=0.19.5
Pillow>=10.0
qrcode>=7.4
pycaw>=20240210
comtypes>=1.3
psutil>=5.9
pyinstaller>=6.0
